refactor(fl): move trace.{h,cpp.hpp} to fl/system/
Relocates fl/trace.{h,cpp.hpp} to fl/system/ alongside log.{h,cpp.hpp}
as siblings of the debug/observability subsystem. Updates unity-build
aggregators and all consumers.
