+++ src/fl/_build.cpp.hpp
@@ -12,7 +12,6 @@
 #include "fl/static_constexpr_defs.cpp.hpp"
 #include "fl/str_ui.cpp.hpp"
 #include "fl/time_alpha.cpp.hpp"
-#include "fl/trace.cpp.hpp"
 
 // begin sub directory includes
 #include "fl/audio/_build.cpp.hpp"
