+++ src/fl/channels/manager.cpp.hpp
@@ -10,7 +10,7 @@
 #include "fl/stl/chrono.h"
 #include "fl/stl/algorithm.h"
 #include "fl/stl/move.h"
-#include "fl/trace.h"
+#include "fl/system/trace.h"
 #include "fl/task/executor.h"
 #include "platforms/init_channel_driver.h"
 #include "fl/stl/noexcept.h"
