+++ src/fl/task/task.h
@@ -104,7 +104,7 @@ auto t = fl::task::coroutine({
 
 #include "fl/stl/functional.h"  // IWYU pragma: keep
 #include "fl/stl/string.h"
-#include "fl/trace.h"
+#include "fl/system/trace.h"
 #include "fl/task/promise.h"  // IWYU pragma: keep
 #include "fl/stl/shared_ptr.h"
 #include "fl/stl/cstddef.h"
