+++ src/platforms/esp/32/drivers/parlio/channel_driver_parlio.cpp.hpp
@@ -25,7 +25,7 @@
 #include "fl/system/log.h"
 #include "fl/system/log.h"
 #include "fl/stl/algorithm.h"
-#include "fl/trace.h"
+#include "fl/system/trace.h"
 #include "fl/stl/assert.h"
 #include "fl/stl/noexcept.h"
