+++ src/platforms/esp/32/drivers/uart/channel_driver_uart.cpp.hpp
@@ -15,7 +15,7 @@
 #include "fl/system/log.h"
 #include "fl/stl/algorithm.h"
 #include "fl/stl/charconv.h"
-#include "fl/trace.h"
+#include "fl/system/trace.h"
 #include "fl/system/log.h"
 #include "fl/stl/noexcept.h"
